Abstract Parser and Ui from Duke
