+++ src/main/java/task/TaskList.java
@@ -1,7 +1,6 @@
 package task;
 
 import exception.DukeException;
-import storage.Storage;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
